Add Spectre.Console for better UI interaction
This commit replaces the standard .NET Console interactions with the Spectre.Console library in order to improve the user interface. The Spectre.Console library provides more advanced features, such as prompts and ASCII art, to create a more engaging user experience. The package references was updated and a "figlet" font file was added to support this new functionality. FSharpPlus package was removed due to lack of utilization in the project.