Refactor variable names and add workspace content and file"
This commit includes refactoring of a function name from 'fixedName' to 'fixName' for consistency with the standard naming convention. The 'fixedName' variable is updated accordingly.

Additionally, a new variable 'workspaceContents' was added to provide the content for a workspace file, which is subsequently written to a new file identified by 'workspaceFile'. The 'startInfo' was updated to point to this new 'workspaceFile' for opening the workspace.

In 'File.fs', a new function 'writeAllText' was added to write all the contents to the provided file path. This function is being used to create the new workspace file.